Revert "rename U238, Pu239 to have correct-er names"
This reverts commit 775745ac58edbf8296646a70164e34463626524b.

+++ src/main/java/com/gregtechceu/gtceu/api/data/chemical/material/Material.java
@@ -1060,7 +1060,7 @@ public class Material implements Comparable<Material> {
             return this;
         }
 
-        public Builder radioactiveHazard(float multiplier) {
+        public Builder radioactiveHazard(int multiplier) {
             properties.setProperty(HAZARD, new HazardProperty(HazardProperty.HazardTrigger.ANY,
                     GTMedicalConditions.CARCINOGEN, multiplier, true));
             return this;
